Module5_Agent_MCP/requirements.txt
streamlit>=1.28.0
openai>=1.12.0
mcp>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0